+++ contactus.php
@@ -34,7 +34,7 @@
                     <a href="../phone-market-revision" class="nav-link">Home</a>
                 </li>
                 <li class="nav-item">
-                    <a href="product.php" class="nav-link">Product</a>
+                    <a href="productlist.php?page=0" class="nav-link">Product</a>
                 </li>
                 <li class="nav-item">
                     <a href="contactus.php" class="nav-link active">Contact us</a>