+++ php/allnotificationlist.php
@@ -15,10 +15,7 @@
     ");
     $preparedata->execute();
     $data = $preparedata->get_result();
-    // $prepdata = $data->fetch_assoc();
     $preparedata->close();
-    // var_dump($prepdata);
-    // // die;
     if(isset($_GET['search'])){
         if($_GET['search'] != null){
             $data = NULL;
@@ -79,7 +76,7 @@
                     <a href="../../phone-market-revision" class="nav-link">Home</a>
                 </li>
                 <li class="nav-item">
-                    <a href="../product.php" class="nav-link">Product</a>
+                    <a href="../productlist.php" class="nav-link">Product</a>
                 </li>
                 <li class="nav-item">
                     <a href="../contactus.php" class="nav-link">Contact us</a>
@@ -206,6 +203,6 @@
     </div>
 </body>
 <script>
-    error_msg(2);
+    error_msg();
 </script>
 </html>