+++ resource/js/main-js-index.js
@@ -24,13 +24,3 @@ $(window).scroll(function () {
 })
 printLetterByLetter("intro", "A Fake Apple Store", 100);
 printLetterByLetter("intro-secondary", "by UnknownRori", 100);
-// review this
-// $(document).ready(function () {
-//     $('.navbar a').on('click', function (e) {
-//         href = $(this).attr('href');
-//         eHref = $(href);
-//         $('html').animate({
-//             scrollTop: eHref.offset().top - 70
-//         }, 1000, 'easeInOutBack')
-//     })
-// });